refactor: Move ConstrainedDecimal test from separate file into test_json
